requirements.txt
asgiref==3.8.1
Django==5.1.2
django-admin==2.0.2
django-decouple==2.1
django-environ==0.11.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
isoweek==1.3.3
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
screen==1.0.1
six==1.16.0
sqlparse==0.5.1
TimeConvert==3.0.13
tzlocal==5.2
xlwt==1.3.0
